fix: [wayland] Window size exceeds the screen after setting the zoom. (#398)
When the window size exceeds the screen, it will automatically maximize under X11 due to the protocol design. However, under Wayland, it will not maximize automatically, leading to the size exceeding the screen. As a workaround, when the window size exceeds the screen, set it to the screen size.

Log: fix bug

Bug: https://pms.uniontech.com/bug-view-305171.html