fix: Abnormal display of the right-click menu on the taskbar
log: The actions in the .desktop file do not match the group names, and the standard [Desktop Action ...] section is missing, causing the Dock in certain environments to fail to display these two items. Therefore, rename the actions to NewWindow and Quake, and add the [Desktop Action NewWindow] and [Desktop Action Quake] sections.

bug: https://pms.uniontech.com/bug-view-331295.html